Re-arrange some project files
