+++ src/main/scala-2.12/asyncstreams/package.scala
@@ -1,9 +1,19 @@
+import cats.Monad
 import cats.mtl.FunctorEmpty
 import cats.mtl.syntax.empty._
+import cats.syntax.applicative._
 
 import scala.language.higherKinds
 
 package object asyncstreams {
+  implicit class AsyncStreamOps[F[+_]: Monad, A](stream: => AsyncStream[F, A]) {
+    def ~::(el: A) = AsyncStream(Step(el, stream).pure[F])
+  }
+
+  implicit class IterableToAS[T](it: Iterable[T]) {
+    def toAS[F[+_]: Monad](implicit methods: ASImpl[F]): AsyncStream[F, T] = methods.fromIterable(it)
+  }
+
   implicit class FunctorWithFilter[F[_] : FunctorEmpty, A](fa: F[A]) {
     def withFilter(f: A => Boolean): F[A] = fa.filter(f)
   }
